+++ pkg/services/object/get.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"context"
 	"crypto/sha256"
-	"encoding/binary"
 	"errors"
 	"fmt"
 	"io"
@@ -13,8 +12,6 @@ import (
 	"github.com/nspcc-dev/neofs-node/internal/protobuf/protoscan"
 	aclsvc "github.com/nspcc-dev/neofs-node/pkg/services/object/acl/v2"
 	apistatus "github.com/nspcc-dev/neofs-sdk-go/client/status"
-	"github.com/nspcc-dev/neofs-sdk-go/object"
-	oid "github.com/nspcc-dev/neofs-sdk-go/object/id"
 	protoobject "github.com/nspcc-dev/neofs-sdk-go/proto/object"
 	protorefs "github.com/nspcc-dev/neofs-sdk-go/proto/refs"
 	protostatus "github.com/nspcc-dev/neofs-sdk-go/proto/status"
@@ -268,109 +265,10 @@ func (x *getProxyContext) handleChunkResponse(streamProg *getStreamProgress, res
 		}
 	}
 
-	remoteSent := respChunkLen == chunkLen
-	if !remoteSent {
-		if respChunkLen <= maxGetResponseChunkLen {
-			localRespBuf, _ := getBufferForChunkGetResponse()
-
-			chunkBuffers.CopyTo(localRespBuf.SliceBuffer[maxChunkOffsetInGetResponse:])
-
-			bodyf := shiftPayloadChunkInGetResponseBuffer(localRespBuf.SliceBuffer, maxChunkOffsetInGetResponse, respChunkLen)
-
-			if x.respStream.signResponse {
-				n, err := x.respStream.srv.signResponse(localRespBuf.SliceBuffer[bodyf.To:], localRespBuf.SliceBuffer[bodyf.ValueFrom:bodyf.To], nil)
-				if err != nil {
-					return false, fmt.Errorf("sign chunk response: %w", err)
-				}
-				bodyf.To += n
-			}
-
-			localRespBuf.SetBounds(bodyf.From, bodyf.To)
-			respBuf = mem.BufferSlice{localRespBuf}
-		} else {
-			// TODO: in this case we could make respBuf = mem.BufferSlice{prefix, chunkBuffers},
-			//  but then we'd have to provide mem.Buffer from iprotobuf.BuffersSlice
-			bodyFldLen := 1 + protowire.SizeBytes(respChunkLen)
-			fullLen := 1 + protowire.SizeBytes(bodyFldLen)
-			if x.respStream.signResponse {
-				fullLen += maxResponseVerificationHeaderLen
-			}
-
-			b := make(mem.SliceBuffer, fullLen)
-			b[0] = iprotobuf.TagBytes1 // body field
-			off := 1 + binary.PutUvarint(b[1:], uint64(bodyFldLen))
-			b[off] = iprotobuf.TagBytes2 // chunk field
-			off += 1 + binary.PutUvarint(b[off+1:], uint64(respChunkLen))
-			off += chunkBuffers.CopyTo(b[off:])
-			if x.respStream.signResponse {
-				n, err := x.respStream.srv.signResponse(b[off:], b[:off], nil)
-				if err != nil {
-					return false, fmt.Errorf("sign chunk response: %w", err)
-				}
-				b = b[:off+n]
-			}
-
-			respBuf = mem.BufferSlice{b}
-		}
-	}
-
-	if err := x.respStream.base.SendMsg(respBuf); err != nil {
-		return remoteSent, err
-	}
-
-	streamProg.readPayload += chunkLen
-	x.respondedPayload += to - from
-
-	return remoteSent, nil
+	return x.respStream.srv.sendChunkResponse(x.respStream.base, respBuf, chunkBuffers, respChunkLen, chunkLen,
+		x.respStream.signResponse, iprotobuf.TagBytes2, &streamProg.readPayload, &x.respondedPayload, shiftPayloadChunkInGetResponseBuffer)
 }
 
 func (x *getProxyContext) handleSplitInfo(respBuf mem.BufferSlice, buffers iprotobuf.BuffersSlice) (bool, error) {
-	var si object.SplitInfo
-	var opts protoscan.ScanMessageOptions
-
-	compose := !x.req.GetBody().GetRaw()
-	if compose {
-		opts.InterceptBytes = func(num protowire.Number, buffers iprotobuf.BuffersSlice) error {
-			if num == protoobject.FieldSplitInfoSplitID {
-				id := object.NewSplitIDFromV2(buffers.ReadOnlyData())
-				if id == nil {
-					return errors.New("invalid split ID")
-				}
-				si.SetSplitID(id)
-			}
-			return nil
-		}
-		opts.InterceptNested = func(num protowire.Number, buffers iprotobuf.BuffersSlice) error {
-			if num != protoobject.FieldSplitInfoLastPart && num != protoobject.FieldSplitInfoLink && num != protoobject.FieldSplitInfoFirstPart {
-				return protoscan.ErrContinue
-			}
-
-			var opts protoscan.ScanMessageOptions
-			opts.InterceptBytes = func(num2 protowire.Number, buffers iprotobuf.BuffersSlice) error {
-				if num2 == protorefs.FieldObjectIDValue {
-					switch num { //nolint:exhaustive
-					case protoobject.FieldSplitInfoLastPart:
-						si.SetLastPart(oid.ID(buffers.ReadOnlyData()))
-					case protoobject.FieldSplitInfoLink:
-						si.SetLink(oid.ID(buffers.ReadOnlyData()))
-					case protoobject.FieldSplitInfoFirstPart:
-						si.SetFirstPart(oid.ID(buffers.ReadOnlyData()))
-					}
-				}
-				return nil
-			}
-			return protoscan.ScanMessage(buffers, protoscan.ObjectIDScheme, opts)
-		}
-	}
-
-	err := protoscan.ScanMessage(buffers, protoscan.ObjectSplitInfoScheme, opts)
-	if err != nil {
-		return false, fmt.Errorf("handle split info field: %w", err)
-	}
-
-	if compose {
-		return false, object.NewSplitInfoError(&si)
-	}
-
-	return true, x.respStream.base.SendMsg(respBuf)
+	return handleSplitInfo(x.req.GetBody().GetRaw(), x.respStream.base, respBuf, buffers)
 }