+++ pkg/services/object/get/util.go
@@ -189,9 +189,8 @@ func (c *clientWrapper) getObject(exec *execCtx) (*object.Object, io.ReadCloser,
 		return nil, nil, exec.forwardGetRequestFn(exec.ctx, c.client)
 	}
 
-	if exec.isForwardingEnabled() {
-		obj, err := exec.prm.forwarder(exec.ctx, c.client)
-		return obj, nil, err
+	if exec.forwardRangeRequestFn != nil {
+		return nil, nil, exec.forwardRangeRequestFn(exec.ctx, c.client)
 	}
 
 	key, err := exec.key()